c: int64_t

+++ c/day2_test.c
@@ -1,37 +1,38 @@
 #include <assert.h>
+#include <stdint.h>
 
 #include "vm.h"
 
 static void test1() {
-    int data[] = {1, 9, 10, 3, 2, 3, 11, 0, 99, 30, 40, 50};
+    int64_t data[] = {1, 9, 10, 3, 2, 3, 11, 0, 99, 30, 40, 50};
     vm vm = make_vm(data, sizeof(data) / sizeof(data[0]));
     run(&vm);
     assert(vm.mem[0] == 3500);
 }
 
 static void test2() {
-    int data[] = {1, 0, 0, 0, 99};
+    int64_t data[] = {1, 0, 0, 0, 99};
     vm vm = make_vm(data, sizeof(data) / sizeof(data[0]));
     run(&vm);
     assert(vm.mem[0] == 2);
 }
 
 static void test3() {
-    int data[] = {2, 3, 0, 3, 99};
+    int64_t data[] = {2, 3, 0, 3, 99};
     vm vm = make_vm(data, sizeof(data) / sizeof(data[0]));
     run(&vm);
     assert(vm.mem[3] == 6);
 }
 
 static void test4() {
-    int data[] = {2, 4, 4, 5, 99, 0};
+    int64_t data[] = {2, 4, 4, 5, 99, 0};
     vm vm = make_vm(data, sizeof(data) / sizeof(data[0]));
     run(&vm);
     assert(vm.mem[5] == 9801);
 }
 
 static void test5() {
-    int data[] = {1, 1, 1, 4, 99, 5, 6, 0, 99};
+    int64_t data[] = {1, 1, 1, 4, 99, 5, 6, 0, 99};
     vm vm = make_vm(data, sizeof(data) / sizeof(data[0]));
     run(&vm);
     assert(vm.mem[0] == 30);
