+++ c/parse.c
@@ -1,10 +1,11 @@
 #include "parse.h"
 
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 
-int parse_intcode(FILE *f, int result[], int max) {
+int parse_intcode(FILE *f, int64_t result[], int max) {
     int n = 0;
-    while (n < max && fscanf(f, "%d,", &result[n]) != EOF) n++;
+    while (n < max && fscanf(f, "%lld,", &result[n]) != EOF) n++;
     return n;
 }
