+++ c/parse.h
@@ -1,11 +1,12 @@
 #ifndef PARSE_H_
 #define PARSE_H_
 
+#include <stdint.h>
 #include <stdio.h>
 
 // reads a comma-separated list of at most |max| integers from |f| into the
 // array |result|. returns the number of integers read if successful, or -1
 // and sets |errno| otherwise.
-int parse_intcode(FILE* f, int result[], int max);
+int parse_intcode(FILE* f, int64_t result[], int max);
 
 #endif  // PARSE_H_
